Promoting Hive Query Generator v2 to v0
