+++ core/src/main/scala/com/yahoo/maha/core/query/hive/HiveQueryGenerator.scala
@@ -1,24 +1,26 @@
-// Copyright 2017, Yahoo Holdings Inc.
+// Copyright 2018, Oath Holdings Inc.
 // Licensed under the terms of the Apache License 2.0. Please see LICENSE file in project root for terms.
 package com.yahoo.maha.core.query.hive
 
 import com.yahoo.maha.core._
 import com.yahoo.maha.core.dimension._
-import com.yahoo.maha.core.fact.{DerivedFactColumn, _}
+import com.yahoo.maha.core.fact._
 import com.yahoo.maha.core.query._
 import grizzled.slf4j.Logging
 
-import scala.collection.mutable.LinkedHashSet
 import scala.collection.{SortedSet, mutable}
 
 /**
- * Created by shengyao on 12/16/15.
+ * Created by pranavbhole on 10/16/18.
+  * Latest hive query generator with sorting feature
  */
 class HiveQueryGenerator(partitionColumnRenderer:PartitionColumnRenderer, udfStatements: Set[UDFRegistration]) extends HiveQueryGeneratorCommon(partitionColumnRenderer, udfStatements) with Logging {
 
+  override val version = Version.v0
+
   override val engine: Engine = HiveEngine
   override def generate(queryContext: QueryContext): Query = {
-    info(s"Generating Hive query using HiveQueryGeneratorV0, version ${version}")
+    info(s"Generating Hive query using HiveQueryGenerator V0:, version ${version}")
     queryContext match {
       case context : CombinedQueryContext =>
         generateQuery(context)
@@ -34,307 +36,6 @@ class HiveQueryGenerator(partitionColumnRenderer:PartitionColumnRenderer, udfSta
   override def validateEngineConstraints(requestModel: RequestModel): Boolean = {
     requestModel.orFilterMeta.isEmpty
   }
-
- /* private[this] def generateOuterGroupByQuery(queryContext: CombinedQueryContext) : Query = {
-
-    val queryBuilderContext = new QueryBuilderContext
-    val queryBuilder: QueryBuilder = new QueryBuilder(
-      queryContext.requestModel.requestCols.size + 5,
-      queryContext.requestModel.requestSortByCols.size + 1)
-    val requestModel = queryContext.requestModel
-    val publicFact = queryContext.factBestCandidate.publicFact
-    val fact = queryContext.factBestCandidate.fact
-    val factViewAlias = queryBuilderContext.getAliasForTable(fact.name)
-    val requestedCols = queryContext.requestModel.requestCols
-    val columnAliasToColMap = new mutable.HashMap[String, Column]()
-    val columnNamesInOuterGroupBy: LinkedHashSet[String] = new mutable.LinkedHashSet[String]
-    val outerGroupByAlias = "outergroupby"
-
-    def renderOuterColumn(columnInfo: ColumnInfo, queryBuilderContext: QueryBuilderContext, duplicateAliasMapping: Map[String, Set[String]], factCandidate: FactBestCandidate): String = {
-
-      def renderFactCol(alias: String, finalAliasOrExpression: String, col: Column, finalAlias: String): String = {
-        val finalExp = {
-          if (col.isInstanceOf[FactColumn] && col.isDerivedColumn) {
-            val rollupExpression = col.asInstanceOf[DerivedFactColumn].rollupExpression
-            if (rollupExpression.equals(NoopRollup) && columnNamesInOuterGroupBy.contains(finalAlias)) {
-              finalAlias
-            } else {
-              val renderedDerived = col.asInstanceOf[DerivedFactColumn].derivedExpression.render(finalAlias, queryBuilderContext.getColAliasToFactColNameMap).toString
-              renderedDerived
-            }
-          } else {
-            finalAliasOrExpression
-          }
-        }
-        val postFilterAlias = renderNormalOuterColumn(col, finalExp)
-        val finalOuterAlias = renderColumnAlias(alias)
-        queryBuilderContext.setFactColAlias(alias, finalOuterAlias, col)
-        s"""$postFilterAlias $finalOuterAlias"""
-      }
-
-      columnInfo match {
-        case FactColumnInfo(alias) =>
-          val column = fact.columnsByNameMap.get(publicFact.aliasToNameColumnMap.get(alias).get).get
-          if (!queryBuilderContext.containsFactColNameForAlias(alias)) {
-            renderColumnWithAlias(factCandidate.fact, column, alias, Set.empty, true)
-          }
-          QueryGeneratorHelper.handleOuterFactColInfo(queryBuilderContext, alias, factCandidate, renderFactCol, duplicateAliasMapping, factCandidate.fact.name, true)
-        case DimColumnInfo(alias) =>
-          val col = queryBuilderContext.getDimensionColByAlias(alias)
-          val finalAlias = queryBuilderContext.getDimensionColNameForAlias(alias)
-          val referredAlias = s"$outerGroupByAlias.$finalAlias"
-          val postFilterAlias = renderNormalOuterColumn(col, referredAlias)
-          s"""$postFilterAlias $finalAlias"""
-        case ConstantColumnInfo(alias, value) =>
-          val finalAlias = getConstantColAlias(alias)
-          s"""'$value' $finalAlias"""
-        case _ => throw new UnsupportedOperationException("Unsupported Column Type")
-      }
-    }
-
-    def renderRollupExpression(expression: String, rollupExpression: RollupExpression, renderedColExp: Option[String] = None) : String = {
-      rollupExpression match {
-        case SumRollup => s"SUM($expression)"
-        case MaxRollup => s"MAX($expression)"
-        case MinRollup => s"MIN($expression)"
-        case AverageRollup => s"AVG($expression)"
-        case HiveCustomRollup(exp) => {
-          exp.sourceColumns.foreach(col => {
-            renderColumnWithAlias(fact, fact.columnsByNameMap.get(col).get, col, Set.empty, false)
-          })
-          s"(${exp.render(expression, Map.empty, renderedColExp)})"
-        }
-        case NoopRollup => s"($expression)"
-        case any => throw new UnsupportedOperationException(s"Unhandled rollup expression : $any")
-      }
-    }
-
-    def renderDerivedFactCols(derivedCols: List[(Column, String)]):Unit = {
-        val requiredInnerCols: Set[String] =
-          derivedCols.view.map(_._1.asInstanceOf[DerivedColumn]).flatMap(dc => dc.derivedExpression.sourceColumns).toSet
-        derivedCols.foreach {
-          case (column, alias) =>
-            requiredInnerCols.foreach(innerColName => {
-              val innerCol = queryContext.factBestCandidate.fact.columnsByNameMap.get(innerColName).get
-              innerCol match {
-                case d:DimensionColumn =>
-                  renderColumnWithAlias(fact, innerCol, innerColName, Set.empty, false)
-                  queryBuilder.addGroupBy(innerColName)
-                case f:FactColumn =>
-                  if (f.isDerivedColumn) {
-                    renderDerivedFactCols(List((f, f.alias.getOrElse(f.name))))
-                  } else {
-                    val rollupExpression = column.asInstanceOf[FactColumn].rollupExpression
-                    val rollup = rollupExpression match {
-                      case NoopRollup => SumRollup // Need to sum facts in the inner select for outergroupby
-                      case _ => rollupExpression
-                    }
-                    val renderedInnerCol = s"${renderRollupExpression(innerColName, rollup)} $innerColName"
-                    queryBuilder.addFactViewColumn(renderedInnerCol)
-                  }
-              }
-            })
-        }
-    }
-
-    def generateOuterGroupByColumns: (String, String) = {
-      val columnsForGroupBySelect: LinkedHashSet[String] = new mutable.LinkedHashSet[String]
-      val columnsForGroupBy: LinkedHashSet[String] = new mutable.LinkedHashSet[String]
-
-      def addDimGroupByColumns(colNameOrAlias: String, isFromFact: Boolean, queryBuilderContext: QueryBuilderContext) = {
-        val referredAlias = {
-          if (isFromFact) {
-            if (queryBuilderContext.containsFactColNameForAlias(colNameOrAlias)) {
-              s"${queryBuilderContext.getAliasForTable(fact.name)}.${queryBuilderContext.getFactColNameForAlias(colNameOrAlias)}"
-            } else {
-              // It's not a requested column
-              s"${queryBuilderContext.getAliasForTable(fact.name)}.$colNameOrAlias"
-            }
-          } else {
-            val finalAlias = queryBuilderContext.getDimensionColNameForAlias(colNameOrAlias)
-            val tableName = queryBuilderContext.getDimensionForColAlias(colNameOrAlias).name
-            s"${queryBuilderContext.getAliasForTable(tableName)}.$finalAlias"
-          }
-        }
-        columnsForGroupBySelect += referredAlias
-        columnsForGroupBy += referredAlias
-        columnNamesInOuterGroupBy += referredAlias
-      }
-
-      def addDerivedFactColOuterGroupBy(alias: String, col: FactColumn): Unit = {
-        if (col.asInstanceOf[DerivedFactColumn].derivedExpression.expression.hasRollupExpression) {
-          // This is a derived column with aggregate expression (eg UDAF)
-          renderColumnWithAlias(fact, col, alias, Set.empty, false)
-          val finalAlias = renderColumnAlias(col.name)
-          val renderedDerived = s"${queryBuilderContext.getFactColExpressionOrNameForAlias(alias)} $finalAlias"
-          columnsForGroupBySelect.add(renderedDerived)
-          columnNamesInOuterGroupBy += finalAlias
-        } else {
-          val sourceColumnNames = col.asInstanceOf[DerivedFactColumn].derivedExpression.sourceColumns
-          sourceColumnNames.foreach(colName => {
-            val sourceCol = queryContext.factBestCandidate.fact.columnsByNameMap.get(colName).get
-            sourceCol match {
-              case d: DimensionColumn =>
-                addDimGroupByColumns(sourceCol.alias.getOrElse(sourceCol.name), true, queryBuilderContext)
-              case f: FactColumn =>
-                val rollupExpression = sourceCol.asInstanceOf[FactColumn].rollupExpression
-                if (sourceCol.isDerivedColumn) {
-                  if (rollupExpression.equals(NoopRollup)) { // Noop rollups should be expanded in outergroupby clause
-                    renderColumnWithAlias(fact, sourceCol, alias, Set.empty, false)
-                    val finalAlias = renderColumnAlias(sourceCol.name)
-                    val renderedDerived = s"${queryBuilderContext.getFactColExpressionOrNameForAlias(alias)} $finalAlias"
-                    columnsForGroupBySelect.add(renderedDerived)
-                    columnNamesInOuterGroupBy += finalAlias
-                  } else {
-                    addDerivedFactColOuterGroupBy(alias, sourceCol.asInstanceOf[DerivedFactColumn])
-                  }
-                } else {
-                  val finalAlias = colName
-                  val rollup = rollupExpression match {
-                    case NoopRollup => SumRollup // Need to sum facts in the groupby
-                    case _ => rollupExpression
-                  }
-                  val rollupExp = s"${renderRollupExpression(finalAlias, rollup)} $finalAlias"
-                  columnsForGroupBySelect += rollupExp
-                  columnNamesInOuterGroupBy += finalAlias
-                }
-            }
-          })
-        }
-      }
-
-      requestedCols.foreach {
-        case FactColumnInfo(alias) =>
-          if (queryBuilderContext.containsFactAliasToColumnMap(alias) && queryBuilderContext.getFactColByAlias(alias).isInstanceOf[DimensionColumn]) {
-            addDimGroupByColumns(alias, true, queryBuilderContext)
-          } else {
-            val colName = queryContext.factBestCandidate.publicFact.aliasToNameColumnMap.get(alias).get
-            val col = queryContext.factBestCandidate.fact.factColMap.get(colName).get
-            if (col.isDerivedColumn) {
-              addDerivedFactColOuterGroupBy(alias, col)
-            } else {
-              val rollup = queryBuilderContext.getFactColByAlias(alias).asInstanceOf[FactColumn].rollupExpression
-              val finalAlias = queryBuilderContext.getFactColExpressionOrNameForAlias(alias)
-              val rollupExp = s"${renderRollupExpression(finalAlias, rollup)} $finalAlias"
-              columnsForGroupBySelect += rollupExp
-              columnNamesInOuterGroupBy += finalAlias
-            }
-          }
-        case DimColumnInfo(alias) =>
-          addDimGroupByColumns(alias, false, queryBuilderContext)
-      }
-      (columnsForGroupBySelect.mkString(","), columnsForGroupBy.mkString(","))
-    }
-
-    def renderColumnWithAlias(fact: Fact,
-                              column: Column,
-                              alias: String,
-                              requiredInnerCols: Set[String],
-                              renderOuterColumn: Boolean): Unit = {
-      val name = column.alias.getOrElse(column.name)
-      val exp = column match {
-        case any if queryBuilderContext.containsColByName(name) && queryBuilderContext.aliasColumnMap.contains(alias) =>
-          //do nothing, we've already processed it
-          ""
-        case DimCol(_, dt, _, _, _, _) if dt.hasStaticMapping =>
-          val renderedAlias = if (!renderOuterColumn) renderColumnAlias(name) else renderColumnAlias(alias)
-          queryBuilderContext.setFactColAliasAndExpression(alias, renderedAlias, column, Option(name))
-          s"${renderStaticMappedDimension(column)} $name"
-        case DimCol(_, dt, _, _, _, _) =>
-          val renderedAlias = if (!renderOuterColumn) renderColumnAlias(name) else renderColumnAlias(alias)
-          queryBuilderContext.setFactColAliasAndExpression(alias, renderedAlias, column, Option(name))
-          name
-        case HiveDerDimCol(_, dt, _, de, _, _, _) =>
-          val renderedAlias = if (!renderOuterColumn) renderColumnAlias(name) else renderColumnAlias(alias)
-          queryBuilderContext.setFactColAlias(alias, renderedAlias, column)
-          s"""${de.render(name, Map.empty)} $renderedAlias"""
-        case FactCol(_, dt, _, rollup, _, _, _) =>
-          dt match {
-            case DecType(_, _, Some(default), Some(min), Some(max), _) =>
-              val renderedAlias = if (!renderOuterColumn) renderColumnAlias(name) else renderColumnAlias(alias)
-              val minMaxClause = s"CASE WHEN (($name >= $min) AND ($name <= $max)) THEN $name ELSE $default END"
-              queryBuilderContext.setFactColAlias(alias, renderedAlias, column)
-              s"""${renderRollupExpression(name, rollup, Option(minMaxClause))} $renderedAlias"""
-            case IntType(_, _, Some(default), Some(min), Some(max)) =>
-              val renderedAlias = if (!renderOuterColumn) renderColumnAlias(name) else renderColumnAlias(alias)
-              val minMaxClause = s"CASE WHEN (($name >= $min) AND ($name <= $max)) THEN $name ELSE $default END"
-              queryBuilderContext.setFactColAlias(alias, renderedAlias, column)
-              s"""${renderRollupExpression(name, rollup, Option(minMaxClause))} $renderedAlias"""
-            case _ =>
-              val renderedAlias = if (!renderOuterColumn) renderColumnAlias(name) else renderColumnAlias(alias)
-              queryBuilderContext.setFactColAliasAndExpression(alias, renderedAlias, column, Option(name))
-              s"""${renderRollupExpression(name, rollup)} $name"""
-          }
-        case HiveDerFactCol(_, _, dt, cc, de, annotations, rollup, _)
-          if renderOuterColumn && (queryContext.factBestCandidate.filterCols.contains(name) || de.expression.hasRollupExpression || requiredInnerCols(name)
-            || de.isDimensionDriven) =>
-          val renderedAlias = if (!renderOuterColumn) renderColumnAlias(name) else renderColumnAlias(alias)
-          queryBuilderContext.setFactColAlias(alias, renderedAlias, column)
-          s"""${renderRollupExpression(de.render(name, Map.empty), rollup)} $renderedAlias"""
-
-        case HiveDerFactCol(_, _, dt, cc, de, annotations, _, _) =>
-          //means no fact operation on this column, push expression outside
-          de.sourceColumns.foreach {
-            case src if src != name =>
-              val sourceCol = fact.columnsByNameMap(src)
-              //val renderedAlias = renderColumnAlias(sourceCol.name)
-              val renderedAlias = sourceCol.alias.getOrElse(sourceCol.name)
-              renderColumnWithAlias(fact, sourceCol, renderedAlias, requiredInnerCols, renderOuterColumn)
-            case _ => //do nothing if we reference ourselves
-          }
-          //val renderedAlias = renderColumnAlias(alias)
-          val renderedAlias = if (!renderOuterColumn) renderColumnAlias(name) else renderColumnAlias(alias)
-          queryBuilderContext.setFactColAliasAndExpression(alias, renderedAlias, column, Option(s"""(${de.render(renderedAlias, queryBuilderContext.getColAliasToFactColNameMap, expandDerivedExpression = true)})"""))
-          ""
-      }
-
-      queryBuilder.addFactViewColumn(exp)
-    }
-
-    val factQueryFragment = generateFactQueryFragment(queryContext, queryBuilder, renderDerivedFactCols, renderRollupExpression, renderColumnWithAlias)
-    generateDimSelects(queryContext.dims, queryBuilderContext, queryBuilder, requestModel, fact, factViewAlias)
-    val (columnsForGroupBySelect, columnsForGroupBy) = generateOuterGroupByColumns
-    val outerCols = generateOuterColumns(queryContext, queryBuilderContext, queryBuilder, renderOuterColumn)
-    val concatenatedCols = generateConcatenatedCols(queryContext, queryBuilderContext)
-
-    val parameterizedQuery : String = {
-      val dimJoinQuery = queryBuilder.getJoinExpressions
-      s"""SELECT $concatenatedCols
-         |FROM(
-         |SELECT $outerCols
-         |FROM(
-         |SELECT $columnsForGroupBySelect
-         |FROM($factQueryFragment)
-         |$factViewAlias
-         |$dimJoinQuery
-         |GROUP BY $columnsForGroupBy) $outerGroupByAlias
-         |)
-     """.stripMargin
-    }
-
-    // generate alias for request columns
-    requestedCols.foreach {
-      case FactColumnInfo(alias) =>
-        columnAliasToColMap += queryBuilderContext.getFactColNameForAlias(alias) -> queryBuilderContext.getFactColByAlias(alias)
-      case DimColumnInfo(alias) =>
-        columnAliasToColMap += queryBuilderContext.getDimensionColNameForAlias(alias) -> queryBuilderContext.getDimensionColByAlias(alias)
-      case ConstantColumnInfo(alias, value) =>
-      //TODO: fix this
-      //constantColumnsMap += renderColumnAlias(alias) -> value
-    }
-    val paramBuilder = new QueryParameterBuilder
-
-    new HiveQuery(
-      queryContext,
-      parameterizedQuery,
-      Option(udfStatements),
-      paramBuilder.build(),
-      queryContext.requestModel.requestCols.map(_.alias),
-      columnAliasToColMap.toMap,
-      IndexedSeq.empty
-    )
-  }
-*/
 
   private[this] def generateQuery(queryContext: CombinedQueryContext) : Query = {
 
@@ -356,8 +57,32 @@ class HiveQueryGenerator(partitionColumnRenderer:PartitionColumnRenderer, udfSta
 
     def renderOuterColumn(columnInfo: ColumnInfo, queryBuilderContext: QueryBuilderContext, duplicateAliasMapping: Map[String, Set[String]], factCandidate: FactBestCandidate): String = {
 
+      def renderNormalOuterColumnWithoutCasting(column: Column, finalAlias: String) : String = {
+        val renderedCol = column.dataType match {
+          case DecType(_, _, Some(default), Some(min), Some(max), _) =>
+            val minMaxClause = s"CASE WHEN (($finalAlias >= ${min}) AND ($finalAlias <= ${max})) THEN $finalAlias ELSE ${default} END"
+            s"""ROUND(COALESCE($minMaxClause, ${default}), 10)"""
+          case DecType(_, _, Some(default), _, _, _) =>
+            s"""ROUND(COALESCE($finalAlias, ${default}), 10)"""
+          case DecType(_, _, _, _, _, _) =>
+            s"""ROUND(COALESCE($finalAlias, 0L), 10)"""
+          case IntType(_,sm,_,_,_) =>
+            s"""COALESCE($finalAlias, 0L)"""
+          case DateType(_) => s"""getFormattedDate($finalAlias)"""
+          case StrType(_, sm, df) =>
+            val defaultValue = df.getOrElse("NA")
+            s"""COALESCE($finalAlias, "$defaultValue")"""
+          case _ => s"""COALESCE($finalAlias, "NA")"""
+        }
+        if (column.annotations.contains(EscapingRequired)) {
+          s"""getCsvEscapedString(CAST(NVL($finalAlias, '') AS STRING))"""
+        } else {
+          renderedCol
+        }
+      }
+
       def renderFactCol(alias: String, finalAliasOrExpression: String, col: Column, finalAlias: String): String = {
-        val postFilterAlias = renderNormalOuterColumn(col, finalAliasOrExpression)
+        val postFilterAlias = renderNormalOuterColumnWithoutCasting(col, finalAliasOrExpression)
         s"""$postFilterAlias $finalAlias"""
       }
 
@@ -369,7 +94,7 @@ class HiveQueryGenerator(partitionColumnRenderer:PartitionColumnRenderer, udfSta
           val finalAlias = queryBuilderContext.getDimensionColNameForAlias(alias)
           val publicDim = queryBuilderContext.getDimensionForColAlias(alias)
           val referredAlias = s"${queryBuilderContext.getAliasForTable(publicDim.name)}.$finalAlias"
-          val postFilterAlias = renderNormalOuterColumn(col, referredAlias)
+          val postFilterAlias = renderNormalOuterColumnWithoutCasting(col, referredAlias)
           s"""$postFilterAlias $finalAlias"""
         case ConstantColumnInfo(alias, value) =>
           val finalAlias = getConstantColAlias(alias)
@@ -465,14 +190,51 @@ class HiveQueryGenerator(partitionColumnRenderer:PartitionColumnRenderer, udfSta
       queryBuilder.addFactViewColumn(exp)
     }
 
+    def generateOrderByClause(queryContext: CombinedQueryContext,
+                             queryBuilder: QueryBuilder): Unit = {
+      val model = queryContext.requestModel
+      model.requestSortByCols.map {
+        case FactSortByColumnInfo(alias, order) =>
+          val colExpression = queryBuilderContext.getFactColNameForAlias(alias)
+          s"$colExpression ${order}"
+        case DimSortByColumnInfo(alias, order) =>
+          val dimColName = queryBuilderContext.getDimensionColNameForAlias(alias)
+          s"$dimColName ${order}"
+        case a => throw new IllegalArgumentException(s"Unhandled SortByColumnInfo $a")
+      }.foreach(queryBuilder.addOrderBy(_))
+    }
+
+    def generateConcatenatedColsWithCast(queryContext: QueryContext, queryBuilderContext: QueryBuilderContext): String = {
+      def castAsString(finalAlias: String):String = {
+        s"CAST($finalAlias AS STRING)"
+      }
+      val renderedConcateColumns = queryContext.requestModel.requestCols.map {
+        case ConstantColumnInfo(alias, _) =>
+          val finalAlias = getConstantColAlias(alias)
+          s"""NVL(${castAsString(finalAlias)}, '')"""
+        case DimColumnInfo(alias) =>
+          val finalAlias = queryBuilderContext.getDimensionColNameForAlias(alias)
+          s"""NVL(${castAsString(finalAlias)}, '')"""
+        case FactColumnInfo(alias)=>
+          val finalAlias = queryBuilderContext.getFactColNameForAlias(alias)
+          s"""NVL(${castAsString(finalAlias)}, '')"""
+      }
+      "CONCAT_WS(\",\"," + (renderedConcateColumns).mkString(", ") + ")"
+    }
+
+
     /**
      * Final Query
      */
 
     val factQueryFragment = generateFactQueryFragment(queryContext, queryBuilder, renderDerivedFactCols, renderRollupExpression, renderColumnWithAlias)
     generateDimSelects(dims, queryBuilderContext, queryBuilder, requestModel, fact, factViewAlias)
+
+    generateOrderByClause(queryContext, queryBuilder)
+    val orderByClause = queryBuilder.getOrderByClause
+
     val outerCols = generateOuterColumns(queryContext, queryBuilderContext, queryBuilder, renderOuterColumn)
-    val concatenatedCols = generateConcatenatedCols(queryContext, queryBuilderContext)
+    val concatenatedCols = generateConcatenatedColsWithCast(queryContext, queryBuilderContext)
 
     val parameterizedQuery : String = {
       val dimJoinQuery = queryBuilder.getJoinExpressions
@@ -484,7 +246,8 @@ class HiveQueryGenerator(partitionColumnRenderer:PartitionColumnRenderer, udfSta
           |SELECT $outerCols
           |FROM($factQueryFragment)
           |$factViewAlias
-          |$dimJoinQuery)
+          |$dimJoinQuery
+          |$orderByClause)
        """.stripMargin
     }
 
@@ -507,7 +270,8 @@ class HiveQueryGenerator(partitionColumnRenderer:PartitionColumnRenderer, udfSta
       paramBuilder.build(),
       queryContext.requestModel.requestCols.map(_.alias),
       columnAliasToColMap.toMap,
-      IndexedSeq.empty
+      IndexedSeq.empty,
+      queryGenVersion = Some(this.version)
     )
   }
 }
@@ -516,9 +280,9 @@ object HiveQueryGenerator extends Logging {
   val ANY_PARTITIONING_SCHEME = HivePartitioningScheme("") //no name needed since class name hashcode
 
   def register(queryGeneratorRegistry: QueryGeneratorRegistry, partitionDimensionColumnRenderer:PartitionColumnRenderer, udfStatements: Set[UDFRegistration]) = {
-    if(!queryGeneratorRegistry.isEngineRegistered(HiveEngine, Option(Version.DEFAULT))) {
+    if(!queryGeneratorRegistry.isEngineRegistered(HiveEngine, Option(Version.v0))) {
       val generator = new HiveQueryGenerator(partitionDimensionColumnRenderer:PartitionColumnRenderer, udfStatements)
-      queryGeneratorRegistry.register(HiveEngine, generator)
+      queryGeneratorRegistry.register(HiveEngine, generator, Version.v0)
     } else {
       queryGeneratorRegistry.getDefaultGenerator(HiveEngine).foreach {
         qg =>
@@ -529,3 +293,6 @@ object HiveQueryGenerator extends Logging {
     }
   }
 }
+
+
+
